fixed analyze reflection and appending analyze info to vis.html
